rename the styles folder and push
